traceloop-demo/app/requirements.txt
traceloop-sdk==0.60.0
google-generativeai>=0.8.0
flask==3.1.0
